fix: adjust the config file in linix use cases
