+++ src/pages/privileges/outlets/linixUseCase/config/dataUseCases.config.tsx
@@ -1,0 +1,91 @@
+import { Link } from "react-router-dom";
+import { DetailsModal } from "../components/DetailsModal";
+import { UseCase } from "../types";
+import { MdModeEdit } from "react-icons/md";
+import { Icon } from "@inube/design-system";
+import { DeleteLinixUseCase } from "../delete-linix-use-case";
+import { deleteLinixUseCaseModal } from "../delete-linix-use-case/config/deleteLinixUseCase.config";
+import { deleteItemData } from "@src/mocks/utils/dataMock.service";
+
+const titlesOptions = [
+  {
+    id: "k_Usecase",
+    titleName: "Código",
+    priority: 0,
+  },
+
+  {
+    id: "n_Usecase",
+    titleName: "Nombre",
+    priority: 1,
+  },
+
+  {
+    id: "i_Tipusec",
+    titleName: "Tipo",
+    priority: 2,
+  },
+];
+
+const useCasesBreakPointsConfig = [
+  { breakpoint: "(min-width: 1091px)", totalColumns: 4 },
+  { breakpoint: "(max-width: 1090px)", totalColumns: 3 },
+  { breakpoint: "(max-width: 1002px)", totalColumns: 2 },
+  { breakpoint: "(max-width: 837px)", totalColumns: 3 },
+  { breakpoint: "(max-width: 550px)", totalColumns: 2 },
+  { breakpoint: "(max-width: 360px)", totalColumns: 1 },
+];
+
+export const actionsConfigLinixUseCase = (linixUseCases: UseCase[]) => {
+  const dataDetailsLinixUseCase = (k_Usecase: string) => {
+    const data = [
+      linixUseCases.find(
+        (linixUseCases) => linixUseCases.k_Usecase === k_Usecase
+      )!,
+    ].map((linixUseCasesSelected) => ({
+      Código: linixUseCasesSelected?.k_Usecase,
+      Nombre: linixUseCasesSelected?.n_Usecase,
+      Tipo: linixUseCasesSelected?.i_Tipusec,
+      Descripcion: linixUseCasesSelected?.n_Descrip,
+      "Opción botón cliente servidor":
+        linixUseCasesSelected?.k_Nforma + "-" + linixUseCasesSelected?.k_Ncampo,
+    }));
+    return [...data].shift();
+  };
+
+  const actionsConfig = [
+    {
+      id: "Details",
+      actionName: "Detalles",
+      content: ({ k_Usecase }: { k_Usecase: string }) => (
+        <DetailsModal data={dataDetailsLinixUseCase(k_Usecase)} />
+      ),
+      type: "secondary",
+    },
+    {
+      id: "Edit",
+      actionName: "Editar",
+      content: ({ id }: { id: string }) => (
+        <Link to={`edit/${id}`}>
+          <Icon appearance="dark" cursorHover icon={<MdModeEdit />} />
+        </Link>
+      ),
+      type: "primary",
+    },
+    {
+      id: "Delete",
+      actionName: "Eliminar",
+      content: ({ k_Usecase }: { k_Usecase: string }) => (
+        <DeleteLinixUseCase
+          linixUseCase={k_Usecase}
+          deleteLinixUseCaseModal={deleteLinixUseCaseModal}
+          handleDeleteLinixUseCase={deleteItemData}
+        />
+      ),
+      type: "remove",
+    },
+  ];
+  return actionsConfig;
+};
+
+export { useCasesBreakPointsConfig, titlesOptions };
